zng_wgt_text

Function font_palette_colors

pub fn font_palette_colors(
    child: impl IntoUiNode,
    colors: impl IntoVar<Vec<(u16, Rgba)>>,
) -> UiNode

P Defines custom palette colors that affect Emoji colors.

The palette is selected by font_palette and then each valid index entry in this property replaces the selected color.

Sets the FONT_PALETTE_COLORS_VAR.