zng_wgt_text
Function font_size

pub fn font_size(
    child: impl UiNode,
    size: impl IntoVar<FontSize>,
) -> impl UiNode

P Sets the font size for the widget and descendants.

This property affects all texts inside the widget and the Length::Em unit.

Sets the FONT_SIZE_VAR context var and the LayoutMetrics::font_size.