zng_wgt_text

Function justify_mode

pub fn justify_mode(
    child: impl UiNode,
    mode: impl IntoVar<Justify>,
) -> impl UiNode

P Config the automatic spacing inserted between words and letters when text is aligned to fill.

Text alignment can be set to Align::FILL_X, if that is the case this config is defines how the glyphs are spaced to fill the text block.

Sets the JUSTIFY_MODE_VAR.