zng_wgt_text

Function lang

pub fn lang(child: impl UiNode, lang: impl IntoVar<Langs>) -> impl UiNode

P Sets the text language and script for the widget and descendants.

This property affects all texts inside the widget and the layout direction.

Sets the LANG_VAR and DIRECTION_VAR context vars and the LayoutMetrics::direction. Also sets the access::lang when accessibility is enabled.