zng_wgt_text

Function rich_text

pub fn rich_text(child: impl IntoUiNode, enabled: impl IntoVar<bool>) -> UiNode

P Defines a rich text context.

When enabled text widget descendants of this widget edit the text across the widgets, for example, the caret position moves to the next text widget when it reaches the start/end of the first text widget.

Nested rich text contexts are allowed and are all part of the outermost context, enabling this property in all nested wrap panels is recommended as it enables some rich text behavior like bringing the focused widget to front to avoid clipping the caret.

When enabled and not nested the widget will handle text commands, it also broadcasts focus change events to all text descendants.