Function zng_wgt_text::txt_parse_live

pub fn txt_parse_live(
    child: impl UiNode,
    enabled: impl IntoVar<bool>,
) -> impl UiNode

P If txt_parse tries to parse after any text change immediately.

This is enabled by default, if disabled the PARSE_CMD can be used to update pending parse.

This property sets the TXT_PARSE_LIVE_VAR.