zng_wgt_text
Function underline_position

pub fn underline_position(
    child: impl UiNode,
    position: impl IntoVar<UnderlinePosition>,
) -> impl UiNode

P Defines what font line gets traced by the underline.

By default uses the font configuration, but it usually crosses over glyph descents causing skips on the line, you can set this UnderlinePosition::Descent to fully clear all glyph descents.

Sets the UNDERLINE_POSITION_VAR.