zng_wgt_text
Function white_space

pub fn white_space(
    child: impl UiNode,
    transform: impl IntoVar<WhiteSpace>,
) -> impl UiNode

P Text white space transform.

Can be used to collapse a sequence of spaces into a single one, or to ignore line-breaks. Is WhiteSpace::Preserve by default.

This property is not applied when the text is txt_editable.

Sets the WHITE_SPACE_VAR.