zng_wgt_text

Function with_font_feature

pub fn with_font_feature<C, S, V, D>(
    child: C,
    state: V,
    set_feature: D,
) -> UiNode
where
    C: IntoUiNode,
    S: VarValue,
    V: IntoVar<S>,
    D: FnMut(&mut FontFeatures, S) -> S + Send + 'static,

Include the font feature config in the widget context.

The modifications done in set_feature are visible only in the FONT_FEATURES_VAR in this context, and features already set in a parent context are included.