zng_wgt_text

Function with_font_palette_color

pub fn with_font_palette_color(
    child: impl UiNode,
    index: u16,
    color: impl IntoVar<Rgba>,
) -> impl UiNode

Set the palette color in the font palette colors.

The index is pushed or replaced on the context FONT_PALETTE_COLORS_VAR.

This function is a helper for declaring properties that configure the colors of a specific font, you can use font_palette_colors to set all color overrides directly.