zng_wgt_text
Function with_font_variation

pub fn with_font_variation(
    child: impl UiNode,
    name: FontVariationName,
    value: impl IntoVar<f32>,
) -> impl UiNode

Includes the font variation config in the widget context.

The variation name is set for the FONT_VARIATIONS_VAR in this context, variations already set in the parent context that are not the same name are also included.