zng_wgt_text::icon

Function ico

pub fn ico(wgt: &mut WidgetBuilding, ico: impl IntoVar<GlyphIcon>)

B The glyph icon.

Mixin Property

This property is a mixin, it modifies the widget builder during build, it does not produce a standalone UI node.