Macro zng_wgt_text::Strong

macro_rules! Strong {
    ($txt:expr) => { ... };
    ($txt:tt, $($format:tt)*) => { ... };
}

A simple text run with bold font weight.

The input syntax is the same as the shorthand Text!.

Configure

Apart from the font weight this widget can be configured with contextual properties like Text!.