zng_wgt_text::node
Function resolve_text

pub fn resolve_text(child: impl UiNode, text: impl IntoVar<Txt>) -> impl UiNode

An UI node that resolves the text context vars, applies the text transform and white space correction and segments the text.

This node setups the ResolvedText for all inner nodes, the Text! widget includes this node in the NestGroup::EVENT group, so all properties except NestGroup::CONTEXT have access using the TEXT::resolved method.

This node also sets the accessibility label to the resolved text.