Struct zng_wgt_text::node::TEXT

pub struct TEXT;

Text internals used by text implementer nodes and properties.

The text implementation is split between two contexts, resolve_text and layout_text, this service provides access to data produced y these two contexts.

Implementations

impl TEXT

pub fn try_resolved(&self) -> Option<RwLockReadGuardOwned<ResolvedText>>

Read lock the current contextual resolved text if called in a node inside resolve_text.

Note that this will block until a read lock can be acquired.

pub fn resolved(&self) -> RwLockReadGuardOwned<ResolvedText>

Read lock the current contextual resolved text.

Panics

Panics if requested in a node outside resolve_text.

pub fn try_laidout(&self) -> Option<RwLockReadGuardOwned<LaidoutText>>

Read lock the current contextual laidout text if called in a node inside layout_text.

Note that this will block until a read lock can be acquired.

pub fn laidout(&self) -> RwLockReadGuardOwned<LaidoutText>

Read lock the current contextual laidout text.

Panics

Panics if not available in context. Is only available inside layout_text after the first layout.

pub fn resolve_caret( &self ) -> MappedRwLockWriteGuardOwned<ResolvedText, CaretInfo>

Write lock the current contextual resolved text to edit the caret.

Panics

Panics if requested in a node outside resolve_text.