zng_wgt_text

Struct SelectionToolbarMix

pub struct SelectionToolbarMix<P>(/* private fields */);

m Selection toolbar properties.

Implementations

impl SelectionToolbarMix<()>

pub fn context_vars_set(set: &mut ContextValueSet)

Insert context variables used by properties in this mixin.

impl<P: WidgetImpl> SelectionToolbarMix<P>

pub fn selection_toolbar(&self, toolbar: impl IntoUiNode)

P Defines the floating mini-toolbar that shows near a new text selection.

The toolbar is used

impl<P: WidgetImpl> SelectionToolbarMix<P>

pub fn selection_toolbar_fn( &self, toolbar: impl IntoVar<WidgetFn<SelectionToolbarArgs>>, )

P Defines the floating mini-toolbar that shows near a new text selection.

Sets the SELECTION_TOOLBAR_FN_VAR.

impl<P: WidgetImpl> SelectionToolbarMix<P>

pub fn selection_toolbar_anchor(&self, offset: impl IntoVar<AnchorOffset>)

P Position the selection toolbar in relation to the bounding box of all selection rectangles.

See selection_toolbar_fn.

Sets the SELECTION_TOOLBAR_ANCHOR_VAR.