zng_wgt_text
Module icon

Glyph icon widget, properties and nodes.

Note that no icons are embedded in this crate directly, you can manually create a GlyphIcon or use an icon set crate. See the zng::icon::material module for an example.

Structs

GlyphIcon

Represents an icon glyph and font.

Icon

W Render icons defined as glyphs in an icon font.

Enums

GlyphSource

Identifies an icon glyph in the font set.

Statics

ICON_COLOR_VAR

Defines the color of an icon.

ICON_SIZE_VAR

Defines the size of an icon.

Functions

ico

c The glyph icon.

ico_color

P Sets the icon font color.

ico_size

P Sets the icon font size.