zng_wgt_text

Module node

UI nodes used for building a text widget.

Structs

CaretInfo

Represents the caret position at the ResolvedText context.

ImePreview

IME text edit that is not committed yet.

LaidoutText

Represents the laidout text.

PendingLayout

Text layout parts that need rebuild.

RenderInfo

Info about the last text render or render update.

ResolvedText

Represents the resolved fonts and the transformed, white space corrected and segmented text.

RichCaretInfo

Represents the caret position in a RichText context.

RichLineInfo

Info about how a rich text leaf defines new lines in a rich text.

RichText

Represents the rich text context.

RichTextCopyParam

Parameter used in COPY_CMD requests sent from the rich text context to collect all selected text for copy.

TEXT

Text internals used by text implementer nodes and properties.

Enums

RichTextComponent

Represents what kind of component the widget is in a rich text tree.

SelectionBy

Defines the source of the current selection.

Traits

RichTextWidgetInfoExt

Extends WidgetInfo state to provide information about rich text.

Functions

default_interactive_caret_visual

Default interactive caret visual.

interactive_carets

An Ui node that implements interaction and renders the interactive carets.

layout_text

An UI node that layouts the parent ResolvedText defined by the text context vars.

line_placeholder

Create a node that is sized one text line height by width.

non_interactive_caret

An Ui node that renders the edit caret visual.

render_ime_preview_underlines

An Ui node that renders the default IME preview underline visual using the parent LaidoutText.

render_overlines

An Ui node that renders the default overline visual using the parent LaidoutText.

render_selection

An Ui node that renders the text selection background.

render_strikethroughs

An Ui node that renders the default strikethrough visual using the parent LaidoutText.

render_text

An UI node that renders the parent LaidoutText.

render_underlines

An Ui node that renders the default underline visual using the parent LaidoutText.

resolve_text

An UI node that resolves the text context vars, applies the text transform and white space correction and segments the text.

rich_text_component

An UI node that implements some behavior for rich text composition.

selection_toolbar_node

Implements the selection toolbar.

set_interactive_caret_spot

Set the caret hotspot that marks the middle of the caret on the text line.