Struct zng_wgt_text::FontFeaturesMix

pub struct FontFeaturesMix<P>(/* private fields */);

m Advanced font config, features, kerning, variations and more.

All properties in this mixin affects Text! nodes inside the widget where they are set.

Implementations

impl FontFeaturesMix<()>

pub fn context_vars_set(set: &mut ContextValueSet)

Insert context variables used by properties in this mix-in.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_variations(&self, variations: impl IntoVar<FontVariations>)

P Sets font variations.

Note: This property fully replaces the font variations for the widget and descendants, use with_font_variation to create a property that sets a variation but retains others from the context.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_features(&self, features: impl IntoVar<FontFeatures>)

P Sets font features.

Note: This property fully replaces the font variations for the widget and descendants, use with_font_variation to create a property that sets a variation but retains others from the context.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_kerning(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font kerning feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_common_lig(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font common ligatures features.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_discretionary_lig(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font discretionary ligatures feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_historical_lig(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font historical ligatures feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_contextual_alt(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font contextual alternatives feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_caps(&self, state: impl IntoVar<CapsVariant>)

P Sets the font capital variant features.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_numeric(&self, state: impl IntoVar<NumVariant>)

P Sets the font numeric variant features.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_num_spacing(&self, state: impl IntoVar<NumSpacing>)

P Sets the font numeric spacing features.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_num_fraction(&self, state: impl IntoVar<NumFraction>)

P Sets the font numeric fraction features.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_swash(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font swash features.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_stylistic(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font stylistic alternative feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_historical_forms(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font historical forms alternative feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_ornaments(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font ornaments alternative feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_annotation(&self, state: impl IntoVar<FontFeatureState>)

P Sets the font annotation alternative feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_style_set(&self, state: impl IntoVar<FontStyleSet>)

P Sets the font stylistic set alternative feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_char_variant(&self, state: impl IntoVar<CharVariant>)

P Sets the font character variant alternative feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_position(&self, state: impl IntoVar<FontPosition>)

P Sets the font sub/super script position alternative feature.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_jp_variant(&self, state: impl IntoVar<JpVariant>)

P Sets the Japanese logographic set.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_cn_variant(&self, state: impl IntoVar<CnVariant>)

P Sets the Chinese logographic set.

impl<P: WidgetImpl> FontFeaturesMix<P>

pub fn font_ea_width(&self, state: impl IntoVar<EastAsianWidth>)

P Sets the East Asian figure width.